Re-use JS the old way
