+++ Gemfile
@@ -38,6 +38,7 @@ gem 'show_for'
 gem 'font-awesome-rails'
 gem 'devise'
 gem 'jquery-rails'
+gem 'jquery-ui-rails'
 gem 'coffee-rails'
 gem 'doorkeeper'
 gem 'rspec'
