+++ app/views/layouts/application.html.haml
@@ -23,7 +23,8 @@
       Pakotoa
     %meta{ name: "viewport", content: "width=device-width, initial-scale=1.0" }
     = stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload'
-    = javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' 
+    = javascript_pack_tag 'application', 'data-turbolinks-track': 'reload'
+    = javascript_include_tag 'application', 'data-turbolinks-track': 'reload'
     = csrf_meta_tags
     = csp_meta_tag
   %body
