+++ package.json
@@ -6,8 +6,6 @@
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "4.2.2",
-    "jquery": "^3.4.1",
-    "jquery-ui": "^1.12.1",
     "turbolinks": "^5.2.0",
     "webpack": "^4.41.6",
     "webpack-cli": "^3.3.11"
